Update memory interface to use u32

+++ emu/src/memory.rs
@@ -2,6 +2,7 @@ use std::alloc::{self, Layout};
 use std::ptr::{read_volatile, write_volatile};
 
 pub struct Memory {
+    #[allow(dead_code)]
     size: u32, // size in bytes
     addr: *mut u8,
 }
@@ -13,27 +14,27 @@ impl Memory {
         Memory { size, addr }
     }
 
-    pub fn rb(&self, addr: usize) -> u8 {
-        unsafe { read_volatile((self.addr as usize + addr) as *const u8) }
+    pub fn rb(&self, addr: u32) -> u8 {
+        unsafe { read_volatile((self.addr as usize + addr as usize) as *const u8) }
     }
 
-    pub fn wb(&mut self, addr: usize, value: u8) {
-        unsafe { write_volatile((self.addr as usize + addr) as *mut u8, value) }
+    pub fn wb(&mut self, addr: u32, value: u8) {
+        unsafe { write_volatile((self.addr as usize + addr as usize) as *mut u8, value) }
     }
 
-    pub fn rhw(&self, addr: usize) -> u16 {
-        unsafe { read_volatile((self.addr as usize + addr) as *const u16) }
+    pub fn rhw(&self, addr: u32) -> u16 {
+        unsafe { read_volatile((self.addr as usize + addr as usize) as *const u16) }
     }
 
-    pub fn whw(&mut self, addr: usize, value: u16) {
-        unsafe { write_volatile((self.addr as usize + addr) as *mut u16, value) }
+    pub fn whw(&mut self, addr: u32, value: u16) {
+        unsafe { write_volatile((self.addr as usize + addr as usize) as *mut u16, value) }
     }
 
-    pub fn rw(&self, addr: usize) -> u32 {
-        unsafe { read_volatile((self.addr as usize + addr) as *const u32) }
+    pub fn rw(&self, addr: u32) -> u32 {
+        unsafe { read_volatile((self.addr as usize + addr as usize) as *const u32) }
     }
 
-    pub fn ww(&mut self, addr: usize, value: u32) {
-        unsafe { write_volatile((self.addr as usize + addr) as *mut u32, value) }
+    pub fn ww(&mut self, addr: u32, value: u32) {
+        unsafe { write_volatile((self.addr as usize + addr as usize) as *mut u32, value) }
     }
 }
